docker/geojson-processor/requirements.txt
geopandas==0.14.1
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.0
flask==3.0.0
gunicorn==21.2.0
GDAL==3.6.2
shapely==2.0.2
geojson==3.1.0
python-dotenv==1.0.0
requests==2.31.0
